+++ lib/act-fluent-logger-rails.rb
@@ -1,4 +1,2 @@
 require "act-fluent-logger-rails/version"
 require "act-fluent-logger-rails/logger"
-require 'act-fluent-logger-rails/middleware'
-require "act-fluent-logger-rails/railtie" if defined? Rails