+++ lib/act-fluent-logger-rails/logger.rb
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 require 'fluent-logger'
+require 'active_support/core_ext'
+require 'uri'
+require 'cgi'
 
 module ActFluentLoggerRails
 
@@ -8,56 +11,53 @@ module ActFluentLoggerRails
     # Severity label for logging. (max 5 char)
     SEV_LABEL = %w(DEBUG INFO WARN ERROR FATAL ANY)
 
-    def self.new(config_file: Rails.root.join("config", "fluent-logger.yml"), log_tags: {})
+    def self.new(config_file: Rails.root.join("config", "fluent-logger.yml"),
+                 log_tags: {},
+                 settings: {},
+                 flush_immediately: false)
       Rails.application.config.log_tags = [ ->(request) { request } ] unless log_tags.empty?
-      fluent_config = YAML.load(ERB.new(config_file.read).result)[Rails.env]
-      settings = {
-        tag:  fluent_config['tag'],
-        host: fluent_config['fluent_host'],
-        port: fluent_config['fluent_port'],
-        messages_type: fluent_config['messages_type'],
-      }
+      if (0 == settings.length)
+        fluent_config = if ENV["FLUENTD_URL"]
+                          self.parse_url(ENV["FLUENTD_URL"])
+                        else
+                          YAML.load(ERB.new(config_file.read).result)[Rails.env]
+                        end
+        settings = {
+          tag:  fluent_config['tag'],
+          host: fluent_config['fluent_host'],
+          port: fluent_config['fluent_port'],
+          messages_type: fluent_config['messages_type'],
+          severity_key: fluent_config['severity_key'],
+        }
+      end
+
+      settings[:flush_immediately] ||= flush_immediately
+
       level = SEV_LABEL.index(Rails.application.config.log_level.to_s.upcase)
       logger = ActFluentLoggerRails::FluentLogger.new(settings, level, log_tags)
       logger = ActiveSupport::TaggedLogging.new(logger)
       logger.extend self
     end
 
+    def self.parse_url(fluentd_url)
+      uri = URI.parse fluentd_url
+      params = CGI.parse uri.query
+
+      {
+        fluent_host: uri.host,
+        fluent_port: uri.port,
+        tag: uri.path[1..-1],
+        messages_type: params['messages_type'].try(:first),
+        severity_key: params['severity_key'].try(:first),
+      }.stringify_keys
+    end
+
     def tagged(*tags)
-      default_log = tags[0][0].is_a?(ActionDispatch::Request)
-      new_tags = []
-      if default_log
-        @request = tags[0][0]
-      else
-        new_tags = push_tags(*tags)
-      end
+      @request = tags[0][0]
       yield self
     ensure
-      pop_tags(new_tags.size)
+      flush
     end
-
-    def push_tags(*tags)
-      tags.flatten.reject(&:blank?).tap do |new_tags|
-        @tags.concat new_tags
-      end
-    end
-
-    def pop_tags(size = 1)
-      @tags.pop size
-    end
-
-    def clear_tags!
-      @tags.clear
-    end
-
-    def add_global_data(hsh)
-      @global_log_data.merge(hsh) if hsh.is_a?(Hash)
-    end
-
-    def set_global_data(hsh)
-      @global_log_data = hsh
-    end
-
   end
 
   class FluentLogger < ActiveSupport::Logger
@@ -66,45 +66,45 @@ module ActFluentLoggerRails
       port    = options[:port]
       host    = options[:host]
       @messages_type = (options[:messages_type] || :array).to_sym
-      @tags = []
-      @base_tag = options[:tag]
+      @tag = options[:tag]
+      @severity_key = (options[:severity_key] || :severity).to_sym
+      @flush_immediately = options[:flush_immediately]
       @fluent_logger = ::Fluent::Logger::FluentLogger.new(nil, host: host, port: port)
       @severity = 0
       @messages = []
       @log_tags = log_tags
       @map = {}
-      @formatter = Proc.new {}
-      @global_log_data = {}
-    end
-
-    def current_tag
-      ([@base_tag] + @tags).join('.')
     end
 
     def add(severity, message = nil, progname = nil, &block)
       return true if severity < level
       message = (block_given? ? block.call : progname) if message.blank?
       return true if message.blank?
-      if message.is_a? Hash
-        direct_post(severity, message)
-      else
-        add_message(severity, message)
-      end
+      add_message(severity, message)
       true
-    end
-
-    def direct_post(severity, data)
-      return if !data.is_a?(Hash) || data.empty?
-      @severity = severity if @severity < severity
-      data = {:log => data.merge(@global_log_data)}
-      data = {level: format_severity(@severity), tags: current_tag.split('.'), time: Time.now.to_s}.merge(data)
-      @fluent_logger.post(current_tag, data)
-      @severity = 0
     end
 
     def add_message(severity, message)
       @severity = severity if @severity < severity
-      @messages << message
+
+      message =
+        case message
+        when ::String
+          message
+        when ::Exception
+          "#{ message.message } (#{ message.class })\n" <<
+            (message.backtrace || []).join("\n")
+        else
+          message.inspect
+        end
+
+      if message.encoding == Encoding::UTF_8
+        @messages << message
+      else
+        @messages << message.dup.force_encoding(Encoding::UTF_8)
+      end
+
+      flush if @flush_immediately
     end
 
     def [](key)
@@ -122,15 +122,10 @@ module ActFluentLoggerRails
                  else
                    @messages
                  end
-
-      @tags.push('rails')
-      @map[:level] = format_severity(@severity)
-      @map[:tags] = current_tag.split('.')
-      @map[:time] = Time.now.to_s
-      @map[:log] = {messages: messages}.merge(@global_log_data)
-      rq_i = {}
+      @map[:messages] = messages
+      @map[@severity_key] = format_severity(@severity)
       @log_tags.each do |k, v|
-        rq_i[k] = case v
+        @map[k] = case v
                   when Proc
                     v.call(@request)
                   when Symbol
@@ -139,10 +134,8 @@ module ActFluentLoggerRails
                     v
                   end rescue :error
       end
-      @map[:log][:request_info] = rq_i
-      @fluent_logger.post(current_tag, @map)
+      @fluent_logger.post(@tag, @map)
       @severity = 0
-      @tags.pop
       @messages.clear
       @map.clear
     end