+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.2.0 / Mar 20 2016
+
+* Add severity_key parameter. Is is The key of severity(DEBUG, INFO, WARN, ERROR).
+
+## 0.1.10 / Dec 23 2015
+
+* flush immediately.
+
+## 0.1.9 / Dec 16 2015
+
+ * Added settings: parameter to ActFluentLoggerRails::Logger.new.
+
+## 0.1.8 / Nov 14 2015
+
+ * Output Object#inspect if message is not String and not Exception.
+
+## 0.1.7 / July 30 2015
+
+ * Be able to log exceptions #15.
+
+## 0.1.6 / March 20 2015
+
+ * Fix incompatible character encodings #13.
+
+## 0.1.5 / July 19 2014
+
+ * Fix keynames of EVN['FLUENTD_URL']
+
+## 0.1.4 / July 18 2014
+
+ * Enable to use EVN['FLUENTD_URL']
+
+## 0.1.3 / April 11 2014
+
+ * Rails 4.1.0
+
 ## 0.1.2 / September 30 2013
 
  * Add 'gem.license = "MIT"' to gemspec.